Refactor code handlers
